almost working searchall test in debug mode but not normal run
