+++ server.py
@@ -1,13 +1,10 @@
-# server.py
-
 import tornado.autoreload
 import tornado.ioloop
 import tornado.web
 from tornado.options import define, options
 
-from src.application import Application
 from settings import testing
-
+from src.aswwu import application
 
 if __name__ == "__main__":
     # allow command line arguments e.g. `python server.py --port=8881`
@@ -18,10 +15,8 @@ if __name__ == "__main__":
 
     io_loop = tornado.ioloop.IOLoop.instance()
 
-    application = Application()
-    application.start_server()
-
     # reload the server if changes are made
     if testing['dev']:
         tornado.autoreload.start()
-    io_loop.start()
+
+    server = application.start_server(options.port)
