+++ src/aswwu/default.conf
@@ -1,0 +1,3 @@
+port=8888
+log_name="aswwu"
+current_year="1920"
