+++ tests/default.conf
@@ -1,3 +1,3 @@
 port=8888
-log_file="aswwu"
+log_name="aswwu_test"
 current_year="1920"
